Used double for blinds

+++ lib/models/game/new_game_model.dart
@@ -53,17 +53,17 @@ class NewGameModel {
   String clubCode;
   String title = 'Test';
   GameType gameType = GameType.HOLDEM;
-  int smallBlind = 1;
-  int bigBlind = 2;
-  int straddleBet = 4;
-  int ante = 0;
+  double smallBlind = 1;
+  double bigBlind = 2.0;
+  double straddleBet = 4.0;
+  double ante = 0.0;
   bool utgStraddleAllowed = false;
   int minPlayers = 2;
   int maxPlayers = 9;
   int gameLength = 60;
   bool buyInApproval = true;
-  int rakePercentage = 0;
-  int rakeCap = 0;
+  double rakePercentage = 0;
+  double rakeCap = 0;
   int buyInMin = 30;
   int buyInMax = 100;
   int actionTime = 30;
@@ -197,10 +197,10 @@ class NewGameModel {
 }
 
 class Blinds {
-  int smallBlind = 1;
-  int bigBlind = 2;
-  int straddle = 4;
-  int ante = 0;
+  double smallBlind = 1;
+  double bigBlind = 2;
+  double straddle = 4;
+  double ante = 0;
 
   Blinds({this.smallBlind, this.bigBlind, this.straddle, this.ante});
 }
